requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
sqlparse==0.4.4

# Slack Integration
slack-sdk==3.26.1
slack-bolt==1.18.1

# LLM and AI
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langsmith==0.0.77

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
structlog==23.2.0
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1 